requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
avro==1.11.3
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
executing==2.0.1
fastavro==1.9.4
fasteners==0.19
fastjsonschema==2.19.1
google-api-core==2.18.0
google-apitools==0.5.31
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-bigtable==2.24.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-dlp==3.18.1
google-cloud-language==2.13.4
google-cloud-pubsub==2.22.0
google-cloud-pubsublite==1.11.0
google-cloud-recommendations-ai==0.10.11
google-cloud-resource-manager==1.12.4
google-cloud-spanner==3.47.0
google-cloud-storage==2.16.0
google-cloud-videointelligence==2.13.4
google-cloud-vision==3.7.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio<2
grpcio-status==1.62.1
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
jedi==0.19.1
Jinja2==3.1.3
Js2Py==0.74
jsonpickle==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
numpy==1.24.4
oauth2client==4.1.3
objsize==0.7.0
orjson==3.9.15
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydot==1.4.2
Pygments==2.17.2
pyjsparser==2.7.1
pymongo==4.6.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
setuptools==65.5.0
shapely==2.0.5
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zstandard==0.22.0
